+++ testing/runner/rest/service_contract.go
@@ -1,9 +1,13 @@
 package rest
 
-import "github.com/viant/endly/testing/validator"
+import (
+	"github.com/viant/endly/model"
+	"github.com/viant/endly/testing/validator"
+)
 
 //Request represents a send request
 type Request struct {
+	*model.Repeater
 	URL     string
 	Method  string
 	Request interface{}